Move make_fingerprint function to aleph.text.
