+++ aleph/schema/types.py
@@ -4,7 +4,7 @@ from normality import ascii_text, stringify, collapse_spaces
 from aleph.data.validate import is_partial_date
 from aleph.data.parse import parse_phone, parse_country, parse_email
 from aleph.data.parse import parse_date
-from aleph.data.keys import make_fingerprint
+from aleph.text import make_fingerprint
 from aleph.util import ensure_list
